chore: moves RegioBar to app dir
