+++ packages/core/app/components/region/RegionBar/index.ts
