+++ packages/core/app/components/region/RegionModal/RegionModal.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import {
   Icon,
   RegionModal as UIRegionModal,
