+++ packages/core/app/components/region/RegionModal/section.module.scss
