+++ packages/core/app/components/sections/RegionBar/DefaultComponents.ts
