+++ packages/core/app/components/sections/RegionBar/OverriddenDefaultRegionBar.ts
