+++ packages/core/src/components/navigation/NavbarLinks/NavbarLinks.tsx
@@ -4,11 +4,10 @@ import { NavbarLinksListItem as UINavbarLinksListItem } from '@faststore/ui'
 import type { NavbarLinksProps as UINavbarLinksProps } from '@faststore/ui'
 
 import Link from 'src/components/ui/Link'
-import RegionButton from 'src/components/region/RegionButton'
+import RegionButton from 'app/components/region/RegionButton'
 import type { NavbarProps } from 'src/components/navigation/Navbar'
 
 import { mark } from 'src/sdk/tests/mark'
-import { useOverrideComponents } from 'src/sdk/overrides/OverrideContext'
 
 interface NavbarLinksProps extends UINavbarLinksProps {
   links: NavbarProps['links']
